requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
celery==5.3.6
redis==5.0.1
pandas==2.1.4
python-dotenv==1.0.0
gunicorn==21.2.0
Pillow==10.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
dj-database-url==2.1.0
supabase==2.3.0 